requirements.txt
google-generativeai
flask
flask_session
markdown
python-dotenv
PyMuPDF
gunicorn
pandas
plotly